+++ rollup.conf.js
@@ -1,5 +1,5 @@
 import {babel} from '@rollup/plugin-babel';
-import {terser} from 'rollup-plugin-terser';
+import terser from '@rollup/plugin-terser';
 
 
 export default [