+++ src/index.js
@@ -2,7 +2,7 @@
 // eslint-disable-next-line import/no-unresolved
 import { h as createElement } from 'vue';
 
-/** @type Object{string: Promise<Element>} */
+/** @type Record<string, PromiseWithState<Element>> */
 const cache = {};
 
 /**
@@ -106,7 +106,7 @@ const InlineSvg = {
                 cache[src] = this.download(src);
             }
             // notify svg is unloaded
-            if (this.svgElSource && cache[src].isPending() && !this.keepDuringLoading) {
+            if (this.svgElSource && cache[src].getIsPending() && !this.keepDuringLoading) {
                 this.svgElSource = null;
                 this.$emit('unloaded');
             }
@@ -136,7 +136,7 @@ const InlineSvg = {
         /**
          * Get the contents of the SVG
          * @param {string} url
-         * @returns {Promise<Element>}
+         * @returns {PromiseWithState<Element>}
          */
         download(url) {
             return makePromiseState(new Promise((resolve, reject) => {
@@ -183,23 +183,26 @@ function setTitle(svg, title) {
     } else { // create a title element if one doesn't already exist
         const titleEl = document.createElementNS('http://www.w3.org/2000/svg', 'title');
         titleEl.textContent = title;
-        svg.appendChild(titleEl);
+        // svg.prepend(titleEl);
+        svg.insertBefore(titleEl, svg.firstChild);
     }
 }
 
 /**
- * @typedef {Promise} PromiseWithState
- * @property {Function<boolean>} isPending
+ * @typedef {Promise & object} PromiseWithState
+ * @property {function: boolean} getIsPending
+ * @template T
  */
 
 /**
  * This function allow you to modify a JS Promise by adding some status properties.
- * @param {Promise|PromiseWithState} promise
- * @return {PromiseWithState}
+ * @template {any} T
+ * @param {Promise<T>|PromiseWithState<T>} promise
+ * @return {PromiseWithState<T>}
  */
 function makePromiseState(promise) {
     // Don't modify any promise that has been already modified.
-    if (promise.isPending) return promise;
+    if (promise.getIsPending) return promise;
 
     // Set initial state
     let isPending = true;
@@ -216,7 +219,7 @@ function makePromiseState(promise) {
         },
     );
 
-    result.isPending = function getIsPending() { return isPending; };
+    result.getIsPending = function getIsPending() { return isPending; };
     return result;
 }